rtc 중복제거 해결 완+ 버튼 2번 해결 필요
