+++ src/pages/Room/PDFViewer/PageCanvasGroup/DrawingCanvasController/util.js
@@ -15,7 +15,7 @@ export const getCanvasRef = (drawingCanvasRefs, pageNum, userId) => {
 };
 
 export const imageToCanvas = (canvasImage, canvas, callback) => {
-	console.log("canvas", canvas);
+	// console.log("canvas", canvas);
 	if (canvas) {
 		const context = canvas.getContext("2d");
 		const image = new Image();
